combined existing homepage with new camera code
