+++ pubspec.yaml
@@ -40,6 +40,7 @@ dependencies:
   validators: ^2.0.0+1
   provider: ^6.0.5
   shared_preferences: ^2.0.17
+  image_picker: ^0.8.6+1
 
 dev_dependencies:
   flutter_test:
